requirements.txt
streamlit
PyPDF2
python-docx
beautifulsoup4
requests